% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hvi_score.R
\name{compute_hvi_score}
\alias{compute_hvi_score}
\title{Compute a heat vulnerability index}
\usage{
compute_hvi_score(x)
}
\arguments{
\item{x}{An \code{hvi_municipalities} object.}
}
\value{
An object of class \code{hvi_score}.
}
\description{
Adds a normalised heat vulnerability score (0–1) based on the proportion
of vulnerable population.
}
