% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_municipalities}
\alias{example_municipalities}
\title{Example Spanish municipalities}
\format{
An sf object with 10 rows and 5 columns:
\describe{
\item{codine}{INE municipality code}
\item{NAMEUNIT}{Municipality name}
\item{province_code}{Numeric province code}
\item{Province}{Province name}
\item{geometry}{Municipality polygon geometry}
}
}
\usage{
example_municipalities
}
\description{
A small synthetic sf object with 10 municipalities for illustration.
}
\keyword{datasets}
