% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hvi_municipalities.R
\name{new_hvi_municipalities}
\alias{new_hvi_municipalities}
\title{Create an hvi_municipalities object}
\usage{
new_hvi_municipalities(x)
}
\arguments{
\item{x}{An \code{sf} object produced by \code{join_population_to_municipalities()}.}
}
\value{
An object of class \code{hvi_municipalities}.
}
\description{
Wraps a joined municipalities + population sf object with class metadata.
}
